Move distance metric/measure tests to DistanceMetricSuite or DistanceMeasureSuite
