--- mllib/src/test/scala/org/apache/spark/mllib/linalg/distance/SquaredEuclideanDistanceMeasureSuite.scala
@@ -1,32 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.spark.mllib.linalg.distance
-
-import org.apache.spark.mllib.linalg.Vectors
-
-class SquaredEuclideanDistanceMeasureSuite extends GeneralDistanceMeasureSuite {
-  override def distanceFactory = new SquaredEuclideanDistanceMeasure
-
-  test("the distance should be 45.0") {
-    val v1 = Vectors.dense(2, 3)
-    val v2 = Vectors.dense(5, 9)
-
-    val distance = distanceFactory(v1, v2)
-    val expected = 45.0
-    assert(distance == 45.0, s"the distance should be nearly equal to 45.0, but ${distance}")
-  }
-}
